[trivial] Rename files containing ModelLoader classes to match contents
