+++ tests/unittests/onnxImporterTest.cpp
@@ -17,7 +17,7 @@
 #include "glow/ExecutionEngine/ExecutionEngine.h"
 #include "glow/Graph/Context.h"
 #include "glow/Graph/Graph.h"
-#include "glow/Importer/ONNX.h"
+#include "glow/Importer/ONNXModelLoader.h"
 #include "gtest/gtest.h"
 
 using namespace glow;
