+++ tools/loader/ImageClassifier.cpp
@@ -18,8 +18,8 @@
 
 #include "glow/Base/Image.h"
 #include "glow/Graph/Nodes.h"
-#include "glow/Importer/Caffe2.h"
-#include "glow/Importer/ONNX.h"
+#include "glow/Importer/Caffe2ModelLoader.h"
+#include "glow/Importer/ONNXModelLoader.h"
 
 #include "llvm/ADT/StringSwitch.h"
 #include "llvm/Support/CommandLine.h"
